refactor: Nest library code in libs folder
