+++ lib/README.md
@@ -1,0 +1,3 @@
+## Libs
+
+The libs folder contains modules and files that are auxiliary to the application. Libraries contain utility functions that are non-specific to the application domain. They are used to encapsulate code that is used in multiple places in the application. The libs folder is not a place for domain-specific code.
