+++ lib/context/sigterm.go
@@ -1,4 +1,4 @@
-package ctx
+package context
 
 import (
 	"context"
